backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
asyncpg
redis
python-dotenv
